pyproject.toml
[project]
name = "revocalize"
version = "0.1.0"
description = "Retrieval-based Voice Conversion (RVC) pipeline"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "numpy>=2.3.5",
    "scipy>=1.11.0",
    "torch>=2.9",
    "torchaudio>=2.9.0",
    "torchcodec==0.8",
    "tqdm>=4.66.0",
    "wandb>=0.23.0",
    "matplotlib>=3.10.7",
    "jupyter>=1.1.1",
    "faiss-cpu"
]

[tool.ruff]
line-length = 100
target-version = "py313"
select = [
    "E",
    "F", 
    "I",
    "UP",
    "B",
    "N"
]
exclude = ["__pycache__", ".venv", "build", "dist"]
ignore = ["E501"]